Only load posts by logged in user for Edit Posts
I don't want users to be able to edit posts by other users. I
additionally ended up moving the generation of the authorSlug into
the user Schema, as, again, I think it makes sense to save this immutable
piece of data in the state as opposed to generating it on the fly.