Fix naming and submit btn ux
I prefer handleSubmit as a name because it is a verb describing an
action, which is the pattern for function names that I adhere to. Also,
the form hangs for a tick to let the user know that they succeeded;
during this time, the submit btn should still be disabled, as we don't
want to unnessarily hit the server again.