Close SearchFilter on route change
SearchFilter is a means of finding articles; when you find the article
you are looking for, it should close and present that article to you.
I've moved all the history.listen logic into Navbar because it is a far
simpler implementation (i.e., App.js doesn't have to manage state that
it doesn't care just to pass it to Navbar).