Fix patch post naming error
Previously, if you attempted to patch a post and you didn't alter the
name, the backend would bail on that request. This is because we don't
allow for title duplicates. But that should only apply to new posts and
existing posts in which its title is changed to that of another existing
post.