Fix Title bug in EditPostForm
Prior to this fix, if you changed the title of one of your posts, the
title would be updated, but the app would crash due to a cross-origin
error. This error is because the postSlug is changed and your current
page is a url that doesn't exist anymore. This is why we changed the
edit-posts route to use _id: _id is secure and will never be updated.