Call getPosts at root level
Instead of checking constantly on different pages, it makes more sense
to do it at App.js and not worry about it.

To compensate for this new setup, we had to alter the data type of our
success property to string to prevent collisions that occured in our
AddPostForm. Previously, if you reloaded the page on this component, it
would read the request success from getPosts as a success for
postSubmission, and then naviagte back to home. Strings provide better
clarity and avoid this issue.