Close dropdown on blur events
Prior to this update, if you successfully logged in, the auth dropdown
would still be open. This is bad UX. The user shouldn't have to close it
after the work was done. This update also will close the dropdown on
route changes, and allows the user to close it by clicking/tabbing
outside of the dropdown. These are all good UX options.